# pizzastore/pizzastore/src/main/java/com/example/pizzastore/dto/ToppingDTO.java
package com.example.pizzastore.dto;

import java.math.BigDecimal;

public class ToppingDTO {
    private Long id;
    private String name;
    private BigDecimal price;
    private String description;

    // Constructor
    public ToppingDTO(Long id, String name, BigDecimal price, String description) {
        this.id = id;
        this.name = name;
        this.price = price;
        this.description = description;
    }

    // Getters and Setters
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public BigDecimal getPrice() {
        return price;
    }

    public void setPrice(BigDecimal price) {
        this.price = price;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}